+++ base/number.jl
@@ -150,7 +150,7 @@ copysign(x::Real, y::Real) = ifelse(signbit(x)!=signbit(y), -x, +x)
 
 conj(x::Real) = x
 transpose(x::Number) = x
-ctranspose(x::Number) = conj(x)
+adjoint(x::Number) = conj(x)
 angle(z::Real) = atan2(zero(z), z)
 
 """