+++ base/sparse/sparsevector.jl
@@ -1401,7 +1401,7 @@ vecnorm(x::SparseVectorUnion, p::Real=2) = vecnorm(nonzeros(x), p)
 # Transpose
 # (The only sparse matrix structure in base is CSC, so a one-row sparse matrix is worse than dense)
 @inline transpose(sv::SparseVector) = RowVector(sv)
-@inline ctranspose(sv::SparseVector) = RowVector(conj(sv))
+@inline adjoint(sv::SparseVector) = RowVector(conj(sv))
 
 ### BLAS Level-1