+++ test/distributed_exec.jl
@@ -469,7 +469,7 @@ d[5,1:2:4,8] = 19
 AA = rand(4,2)
 A = @inferred(convert(SharedArray, AA))
 B = @inferred(convert(SharedArray, AA'))
-@test B*A == ctranspose(AA)*AA
+@test B*A == adjoint(AA)*AA
 
 d=SharedArray{Int64,2}((10,10); init = D->fill!(D.loc_subarr_1d, myid()), pids=[id_me, id_other])
 d2 = map(x->1, d)